// Agro-Tech-Angular/src/app/modules/warehouse/models/warehouse.model.ts
export type Warehouse = {
    id?: string
    code?: string
    name?: string
    type?: string
    vendor?: any
    costCenterCode?: string
    costCenterName?: string
    costCenterType?: string|number
    divisionCode?: string
    divisionName?: string
    startingDate?: string
    isPrimary?: boolean
    address1?: string
    address2?: string
    cityCode?: string
    cityName?: string
    wilayaCode?: string
    wilayaName?: string
    zipCode?: string
    email?: string
    phoneNumber?: string
    faxNumber?: string
    latitude?: number
    longitude?: number
    status?: boolean
}